Rewrite README for an updated documentation
Rewrite the documentation in the README file to document the current
behavior and configuration options of Slugify.Core. The previous content
was still mostly based on the original library which no longer matches
the implementation since version 3.0.

Add documentation for the `AllowedChars` option which now drives the
main behavior of the library. Also adjust the documentation of
`DeniedCharactersRegex` to reflect the actual current default (it being
disabled in favor of `AllowedChars`) and describes it as the legacy
behavior which is just kept for compatibility reasons.

Also adjust the documentation of the other options and describe the
default slug generation behavior in text instead of showing the default
configuration object which would be difficult and confusing due to the
internals of the `AllowedChars` set.

- Link NuGet shield to current NuGet release.
- Adjust alternative text for shields.
- Adjust general formatting.
- Remove license information since it is already documented in the
  LICENSE file.